Move llvmlite to llvm19
List of changes in this MR

 - Delete legacy pass manager code
 - Delete typed pointer code
 - Update APIs that needs updating for llvm19
 - Update refprune pass and its tests
