+++ docs/source/user-guide/examples/sum.py
@@ -18,7 +18,6 @@ import llvmlite.ir as ll
 import llvmlite.binding as llvm
 
 
-llvm.initialize()
 llvm.initialize_native_target()
 llvm.initialize_native_asmprinter()
