+++ examples/opaque_pointers/sum.py
@@ -21,7 +21,6 @@ import llvmlite.ir as ll
 import llvmlite.binding as llvm
 
 
-llvm.initialize()
 llvm.initialize_native_target()
 llvm.initialize_native_asmprinter()
