+++ ffi/Makefile.freebsd
@@ -10,7 +10,7 @@ LDFLAGS := $(LDFLAGS) $(LLVM_LDFLAGS) $(LD_FLTO_FLAGS)
 LIBS = $(LLVM_LIBS)
 INCLUDE = core.h
 SRC = assembly.cpp bitcode.cpp core.cpp initfini.cpp module.cpp value.cpp \
-	executionengine.cpp transforms.cpp passmanagers.cpp type.cpp targets.cpp \
+	executionengine.cpp type.cpp targets.cpp \
 	dylib.cpp linker.cpp object_file.cpp orcjit.cpp custom_passes.cpp \
 	memorymanager.cpp newpassmanagers.cpp
 OUTPUT = libllvmlite.so
