+++ ffi/orcjit.cpp
@@ -5,7 +5,7 @@
 
 #include "llvm/AsmParser/Parser.h"
 #include "llvm/ExecutionEngine/Orc/Core.h"
-#include "llvm/ExecutionEngine/Orc/DebuggerSupportPlugin.h"
+#include "llvm/ExecutionEngine/Orc/Debugging/DebuggerSupportPlugin.h"
 #include "llvm/ExecutionEngine/Orc/ExecutionUtils.h"
 #include "llvm/ExecutionEngine/Orc/LLJIT.h"
 #include "llvm/ExecutionEngine/Orc/ObjectLinkingLayer.h"
@@ -92,7 +92,7 @@ LLVMPY_CreateLLJITCompiler(LLVMTargetMachineRef tm, bool suppressErrors,
             auto linkingLayer =
                 std::make_unique<llvm::orc::ObjectLinkingLayer>(session);
 
-            /* TODO(LLVM16): In newer LLVM versions, there is a simple
+            /* FIXME(LLVM16): In newer LLVM versions, there is a simple
              * EnableDebugSupport flag on the builder and we don't need to do
              * any of this. */
             if (triple.getObjectFormat() == Triple::ELF ||
@@ -195,8 +195,11 @@ LLVMPY_LLJIT_Link(std::shared_ptr<LLJIT> *lljit, const char *libraryName,
     for (size_t import_idx = 0; import_idx < imports_length; import_idx++) {
         SymbolStringPtr mangled =
             (*lljit)->mangleAndIntern(imports[import_idx].name);
-        JITEvaluatedSymbol symbol(imports[import_idx].address,
-                                  JITSymbolFlags::Exported);
+        ExecutorSymbolDef symbol(ExecutorAddr(imports[import_idx].address),
+                                 JITSymbolFlags::Exported);
+
+        llvm::orc::SymbolMap sym_map;
+        sym_map[mangled] = symbol;
         auto error = dylib->define(absoluteSymbols({{mangled, symbol}}));
 
         if (error) {
