+++ ffi/targets.cpp
@@ -1,14 +1,13 @@
 #include "core.h"
 #include "llvm-c/Target.h"
 #include "llvm-c/TargetMachine.h"
-#include "llvm/ADT/Optional.h"
-#include "llvm/ADT/Triple.h"
 #include "llvm/Analysis/TargetLibraryInfo.h"
 #include "llvm/IR/LegacyPassManager.h"
 #include "llvm/IR/Type.h"
 #include "llvm/MC/TargetRegistry.h"
-#include "llvm/Support/Host.h"
 #include "llvm/Target/TargetMachine.h"
+#include "llvm/TargetParser/Host.h"
+#include "llvm/TargetParser/Triple.h"
 
 #include <cstdio>
 #include <cstring>
@@ -60,9 +59,10 @@ LLVMPY_GetTripleParts(const char *triple_str, const char **arch_out,
  */
 API_EXPORT(int)
 LLVMPY_GetHostCPUFeatures(const char **Out) {
-    llvm::StringMap<bool> features;
+    // https://github.com/llvm/llvm-project/pull/97824
+    llvm::StringMap<bool> features = llvm::sys::getHostCPUFeatures();
     std::ostringstream buf;
-    if (llvm::sys::getHostCPUFeatures(features)) {
+    if (!features.empty()) {
         for (auto &F : features) {
             if (buf.tellp()) {
                 buf << ',';
@@ -147,20 +147,22 @@ LLVMPY_CreateTargetMachine(LLVMTargetRef T, const char *Triple, const char *CPU,
                            const char *RelocModel, const char *CodeModel,
                            int PrintMC, int JIT, const char *ABIName) {
     using namespace llvm;
-    CodeGenOpt::Level cgol;
+
+    // https://github.com/llvm/llvm-project/pull/66295
+    CodeGenOptLevel cgol;
     switch (OptLevel) {
     case 0:
-        cgol = CodeGenOpt::None;
+        cgol = CodeGenOptLevel::None;
         break;
     case 1:
-        cgol = CodeGenOpt::Less;
+        cgol = CodeGenOptLevel::Less;
         break;
     case 3:
-        cgol = CodeGenOpt::Aggressive;
+        cgol = CodeGenOptLevel::Aggressive;
         break;
     case 2:
     default:
-        cgol = CodeGenOpt::Default;
+        cgol = CodeGenOptLevel::Default;
     }
 
     CodeModel::Model cm;
@@ -184,7 +186,8 @@ LLVMPY_CreateTargetMachine(LLVMTargetRef T, const char *Triple, const char *CPU,
             cm = CodeModel::Large;
     }
 
-    Optional<Reloc::Model> rm;
+    // llvm::Optional removed in llvm17
+    std::optional<Reloc::Model> rm;
     std::string rms(RelocModel);
     if (rms == "static")
         rm = Reloc::Static;
