+++ ffi/value.cpp
@@ -6,6 +6,7 @@
 
 // the following is needed for WriteGraph()
 #include "llvm/Analysis/CFGPrinter.h"
+#include "llvm/IR/GlobalVariable.h"
 #include "llvm/Support/GraphWriter.h"
 
 /* An iterator around a attribute list, including the stop condition */
