+++ llvmlite/binding/__init__.py
@@ -8,9 +8,7 @@ from .linker import *
 from .module import *
 from .options import *
 from .newpassmanagers import *
-from .passmanagers import *
 from .targets import *
-from .transforms import *
 from .value import *
 from .typeref import *
 from .analysis import *
